requirements.txt
numpy == 1.24.4
pillow==10.0.0
pillow_heif == 0.13.0
pillow-avif-plugin==1.3.1
pillow-heif==0.13.0
avif==0.5.0
imageio==2.31.1
scikit-image == 0.21.0
pyheif == 0.7.1
ipython == 8.1.0
flask == 2.3.2